Extract method: GCloud.ConfigureGcloudAccount
